+++ packages/config/src/api/site_info.ts
@@ -1,6 +1,5 @@
-import { NetlifyAPI } from '@netlify/api'
-
 import * as z from 'zod'
+import type { NetlifyAPI } from '@netlify/api'
 
 import { getEnvelope } from '../env/envelope.js'
 import { throwUserError } from '../error.js'
@@ -29,8 +28,8 @@ type GetSiteInfoOptions = {
 }
 
 export type Extension = {
-  author: string | undefined
-  extension_token: string | undefined
+  author?: string | undefined
+  extension_token?: string | undefined
   has_build: boolean
   name: string
   slug: string
@@ -172,8 +171,8 @@ const ExtensionResponseSchema = z.array(
     // path. They might call the Netlify API expecting to have an API token available to them when
     // they really don't. For the time being, I've added instrumentation to Jigsaw to help us figure
     // out if this is dead code or actually supports current users.
-    author: z.string().optional().default(undefined),
-    extension_token: z.string().optional().default(undefined),
+    author: z.string().optional(),
+    extension_token: z.string().optional(),
     has_build: z.boolean(),
     name: z.string(),
     slug: z.string(),