+++ packages/config/src/mutations/config_prop_name.ts
@@ -2,12 +2,12 @@
 const DYNAMIC_OBJECT_PROPS = new Set(['build.services', 'build.environment', 'functions', 'functions.*'])
 
 /** Retrieve normalized property name */
-export const getPropName = function (keys: string[]) {
+export const getPropName = function (keys: (string | number)[]) {
   // Some properties are user-defined, i.e. we need to replace them with a "*" token
   // Check if a property name is dynamic, such as `functions.{functionName}`, or
   // is an array index.
   // In those cases, we replace it by "*".
-  return keys.reduce((propName, key) => {
+  return keys.reduce<string>((propName, key) => {
     const normalizedKey = Number.isInteger(key) || DYNAMIC_OBJECT_PROPS.has(propName) ? '*' : String(key)
     return propName === '' ? normalizedKey : `${propName}.${normalizedKey}`
   }, '')