+++ packages/edge-bundler/node/config.ts
@@ -27,7 +27,7 @@ export const enum Cache {
   Manual = 'manual',
 }
 
-export type HTTPMethod = 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE' | 'OPTIONS'
+export type HTTPMethod = 'GET' | 'HEAD' | 'POST' | 'PUT' | 'PATCH' | 'DELETE' | 'OPTIONS'
 
 export type Path = `/${string}`