+++ packages/zip-it-and-ship-it/src/utils/format_result.ts
@@ -7,15 +7,21 @@ import type { ExtendedRoute, Route } from './routes.js'
 export type FunctionResult = Omit<FunctionArchive, 'runtime'> & {
   bootstrapVersion?: string
   eventSubscriptions?: string[]
+  memory?: number
+  region?: string
   routes?: ExtendedRoute[]
   excludedRoutes?: Route[]
   runtime: RuntimeName
   schedule?: string
   runtimeAPIVersion?: number
+  vcpu?: number
 }
 
 // Takes the result of zipping a function and formats it for output.
 export const formatZipResult = (archive: FunctionArchive) => {
+  const memory: number | undefined = archive.staticAnalysisResult?.config?.memory ?? archive?.config?.memory
+  const vcpu: number | undefined = archive.staticAnalysisResult?.config?.vcpu ?? archive?.config?.vcpu
+
   const functionResult: FunctionResult = {
     ...archive,
     staticAnalysisResult: undefined,
@@ -23,8 +29,11 @@ export const formatZipResult = (archive: FunctionArchive) => {
     routes: archive.staticAnalysisResult?.routes,
     excludedRoutes: archive.staticAnalysisResult?.excludedRoutes,
     runtime: archive.runtime.name,
+    memory,
+    region: archive.staticAnalysisResult?.config?.region ?? archive?.config?.region,
     schedule: archive.staticAnalysisResult?.config?.schedule ?? archive?.config?.schedule,
     runtimeAPIVersion: archive.staticAnalysisResult?.runtimeAPIVersion,
+    vcpu,
   }
 
   return removeUndefined(functionResult)